+++ tests/testthat/helper-models.R
@@ -1,10 +1,10 @@
 orthodont_data <- function() {
   data <- as.data.frame(nlme::Orthodont)
-  data$Occasion <- ave(
+  data$Occasion <- factor(ave(
     seq_len(nrow(data)),
     data$Subject,
     FUN = seq_along
-  )
+  ))
   data
 }
 
@@ -22,7 +22,7 @@ fit_multicentre_ar1 <- function() {
     data = multicentre,
     formula = Y ~ Drug * Time,
     random = ~ 1 | Center,
-    repeated = ~ Time | Center:Subject,
+    repeated = ~ Time | Center:Drug:Subject,
     structure = "ar1"
   )
 }